chore: Models directory
