+++ go.mod
@@ -1,0 +1,3 @@
+module mediasms
+
+go 1.15
