+++ models/apiResponse.go
@@ -1,0 +1,8 @@
+package models
+
+// apiResponse struct
+type apiResponse struct {
+	StatusCode  string
+	Name        string
+	Description string
+}
