+++ models/client.go
@@ -1,0 +1,8 @@
+package models
+
+// Client struct
+type Client struct {
+	Username string
+	Password string
+	Prefix   string
+}
