+++ models/webHook.go
@@ -1,0 +1,13 @@
+package models
+
+// WebHook from media4u
+type WebHook struct {
+	Mobilenumber      string
+	Status            string
+	Smsid             string
+	Returnsms         string
+	Raitreturnsms     string
+	Returnsmsdatetime string
+	Replyid           string
+	Senderid          string
+}
